refactor(app): extract agentdir package, split app.go helpers into sibling files
app.go had grown into a mixed composition root holding unrelated helpers.
Move openBrowser, seedSoundsDir, and writeStateFile into focused sibling
files (browser.go, sounds.go, state_file.go) within package app.

Introduce internal/agentdir as the single source of truth for resolving the
Pi agent directory. Previously app.piAgentDir (which respects
PI_CODING_AGENT_DIR) and a weaker HOME-only fallback in server.New computed
the same path two different ways. Both now call agentdir.Path(), and the
repeated filepath.Join(agentDir, "pi-web") convention across app and server
goes through agentdir.WebDir().

agentdir.Path() documents that it intentionally models only the agent-dir
override, mirroring pi's getAgentDir(); the more specific session-dir knobs
(--session-dir, PI_CODING_AGENT_SESSION_DIR) are a known, accepted gap.

Tests: new agentdir package tests (Path env override + fallback, WebDir);
add seedSoundsDir tests (seeds when empty, skips when an mp3 is present);
rename pi_agent_dir_test.go -> state_file_migration_test.go to sit with the
state-file code now that the piAgentDir tests moved to agentdir.
