+++ .gitignore
@@ -13,3 +13,4 @@ node_modules/
 tmp/
 
 .requirements/
+.booboo/
